# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: HireFlux
  description: "\n# HireFlux API\n\nAI-Powered Job Application Copilot that streamlines\
    \ job search with:\n- **Tailored Resumes**: ATS-optimized resumes with multiple\
    \ versions\n- **AI Cover Letters**: Personalized cover letters in formal, concise,\
    \ or conversational tones\n- **Smart Job Matching**: Embeddings-based matching\
    \ with 0-100 Fit Index\n- **Application Tracking**: Comprehensive pipeline management\n\
    - **Auto-Apply**: Consent-based automated job applications\n- **Interview Coach**:\
    \ Mock interviews with AI feedback\n\n## Authentication\n\nAll endpoints (except\
    \ `/auth/*` and `/health`) require Bearer token authentication:\n\n```\nAuthorization:\
    \ Bearer <your_jwt_token>\n```\n\nObtain tokens via:\n- Email/Password: `POST\
    \ /auth/login`\n- Google OAuth: `GET /auth/google/authorize`\n- LinkedIn OAuth:\
    \ `GET /auth/linkedin/authorize`\n\n## Rate Limiting\n\n- **Per Minute**: 60 requests\n\
    - **Per Hour**: 1000 requests\n\n## Pagination\n\nList endpoints support pagination\
    \ via query parameters:\n- `page`: Page number (default: 1)\n- `limit`: Items\
    \ per page (default: 20, max: 100)\n\n## Response Format\n\n### Success Response\n\
    ```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"meta\": {\n    \"\
    page\": 1,\n    \"limit\": 20,\n    \"total\": 100\n  }\n}\n```\n\n### Error Response\n\
...
  contact:
    name: HireFlux Support
...
  license:
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      description: 'Register a new user account.


        Creates a new user with:

        - User account with hashed password

        - User profile

        - Free subscription

        - Credit wallet (0 balance)


        Returns user info and JWT tokens.'
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Register Api V1 Auth Register Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: 'Authenticate user and return JWT tokens.


        Returns:

        - Access token (expires in 24 hours by default)

        - Refresh token (expires in 30 days by default)'
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Login Api V1 Auth Login Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get Current User Info
      description: 'Get current authenticated user information.


        Requires valid JWT token in Authorization header.'
      operationId: get_current_user_info_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Current User Info Api V1 Auth Me Get
      security:
      - HTTPBearer: []
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh Token
      description: 'Refresh access token using refresh token.


        Requires valid refresh token in Authorization header.

        Returns new access and refresh tokens.'
      operationId: refresh_token_api_v1_auth_refresh_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Refresh Token Api V1 Auth Refresh Post
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    UserCreate:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
        terms_accepted:
          type: boolean
          title: Terms Accepted
      type: object
      required:
      - name
      - email
      - password
      - terms_accepted
      title: UserCreate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...